requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
psycopg2==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
jira==3.5.1
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
defusedxml==0.7.1
Flask-APScheduler==1.13.1